:truck: devide helper function in table pages.
