+++ packages/playground-react/helpers/table/index.ts
@@ -1,0 +1,11 @@
+export const getDateForm = (date: number) => {
+  const dateArray = [
+    new Date(date).getFullYear(),
+    new Date(date).getMonth() + 1,
+    new Date(date).getDate(),
+  ];
+
+  return dateArray
+    .map((number) => (number < 10 ? `0${number}` : number))
+    .join('.');
+};
