+++ packages/playground-react/tsconfig.path.json
@@ -6,7 +6,8 @@
       "@pages/*": ["pages/*"],
       "@styles/*": ["styles/*"],
       "@utils/*": ["utils/*"],
-      "@constants/*": ["constants/*"]
+      "@constants/*": ["constants/*"],
+      "@helpers/*": ["helpers/*"]
     }
   }
 }
